requirements-new.txt
absl-py==2.1.0
accelerate==0.25.0
aiohttp==3.9.5
aiosignal==1.3.1
altair==4.2.2
annotated-types==0.7.0
async-timeout==4.0.3
attrs==23.2.0
bitsandbytes==0.43.0
blinker==1.8.2
boto3==1.34.150
botocore==1.34.150
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
diffusers==0.25.0
easygui==0.98.3
einops==0.7.0
entrypoints==0.4
filelock==3.13.1
Flask==3.0.3
frozenlist==1.4.1
fsspec==2024.2.0
ftfy==6.1.1
grpcio==1.65.1
huggingface-hub==0.20.1
idna==3.4
imagesize==1.4.1
importlib_metadata==8.2.0
itsdangerous==2.2.0
Jinja2==3.1.3
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
-e git+https://github.com/kohya-ss/sd-scripts.git@90b7f2f9cafd419b9e728cda6e4e498db3c0739f#egg=library
lightning-utilities==0.11.6
lion-pytorch==0.0.6
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
networkx==3.2.1
numpy==1.26.3
opencv-python==4.7.0.68
packaging==24.1
pandas==2.2.2
pillow==10.2.0
prodigyopt==1.0
protobuf==4.25.4
psutil==6.0.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytorch-lightning==1.9.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.7.24
requests==2.28.1
rich==13.7.0
rpds-py==0.19.1
s3transfer==0.10.2
safetensors==0.4.2
six==1.16.0
sympy==1.12
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.1.2+cu118
torchmetrics==1.4.0.post0
torchvision==0.16.2+cu118
tqdm==4.66.4
transformers==4.36.2
triton==2.1.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==1.26.13
voluptuous==0.13.1
wcwidth==0.2.13
Werkzeug==3.0.3
xformers==0.0.23.post1+cu118
yarl==1.9.4
zipp==3.19.2
